+++ app/src/main/java/com/sopt/geonppang/domain/model/MyReviewDetail.kt
@@ -1,0 +1,7 @@
+package com.sopt.geonppang.domain.model
+
+data class MyReviewDetail(
+    val recommendKeywordList: List<DetailReview.RecommendKeyword>,
+    val reviewText: String,
+    val isLike: Boolean
+)
