+++ app/src/main/java/com/sopt/geonppang/presentation/reviewWriting/ReviewCancelBottomDialogFragment.kt
@@ -1,16 +1,16 @@
-package com.sopt.geonppang.presentation.review
+package com.sopt.geonppang.presentation.reviewWriting
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
 import com.sopt.geonppang.R
-import com.sopt.geonppang.databinding.DialogBottomReviewWritingSuccessBinding
+import com.sopt.geonppang.databinding.DialogBottomReviewWritingCancelBinding
 import com.sopt.geonppang.util.binding.BindingBottomSheetDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class ReviewSuccessBottomDialogFragment :
-    BindingBottomSheetDialogFragment<DialogBottomReviewWritingSuccessBinding>(R.layout.dialog_bottom_review_writing_success) {
+class ReviewCancelBottomDialogFragment :
+    BindingBottomSheetDialogFragment<DialogBottomReviewWritingCancelBinding>(R.layout.dialog_bottom_review_writing_cancel) {
     private val viewModel: ReviewWritingViewModel by activityViewModels()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -21,9 +21,12 @@ class ReviewSuccessBottomDialogFragment :
     }
 
     private fun addListeners() {
-        binding.btnReviewWritingSuccess.setOnClickListener {
-            viewModel.writeReview()
+        binding.tvContinue.setOnClickListener {
             dismiss()
+        }
+
+        binding.tvStop.setOnClickListener {
+            viewModel.setReviewCancelState(false)
         }
     }
 }
