+++ app/src/main/java/com/sopt/geonppang/presentation/reviewWriting/ReviewSuccessBottomDialogFragment.kt
@@ -1,16 +1,16 @@
-package com.sopt.geonppang.presentation.review
+package com.sopt.geonppang.presentation.reviewWriting
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
 import com.sopt.geonppang.R
-import com.sopt.geonppang.databinding.DialogBottomReviewWritingCancelBinding
+import com.sopt.geonppang.databinding.DialogBottomReviewWritingSuccessBinding
 import com.sopt.geonppang.util.binding.BindingBottomSheetDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class ReviewCancelBottomDialogFragment :
-    BindingBottomSheetDialogFragment<DialogBottomReviewWritingCancelBinding>(R.layout.dialog_bottom_review_writing_cancel) {
+class ReviewSuccessBottomDialogFragment :
+    BindingBottomSheetDialogFragment<DialogBottomReviewWritingSuccessBinding>(R.layout.dialog_bottom_review_writing_success) {
     private val viewModel: ReviewWritingViewModel by activityViewModels()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -21,12 +21,9 @@ class ReviewCancelBottomDialogFragment :
     }
 
     private fun addListeners() {
-        binding.tvContinue.setOnClickListener {
+        binding.btnReviewWritingSuccess.setOnClickListener {
+            viewModel.writeReview()
             dismiss()
-        }
-
-        binding.tvStop.setOnClickListener {
-            viewModel.setReviewCancelState(false)
         }
     }
 }
